# Conflicts:
#	.vs/ConsoleApplication2/v14/.suo
#	.vs/PoeProjects/v14/.suo
#	ItemWatcher2/Form1.cs
#	ItemWatcher2/NinjaPoETradeMethods.cs
#	ItemWatcher2/bin/Debug/AllBaseTypesStrings.json
#	ItemWatcher2/bin/Debug/Config.json
#	ItemWatcher2/bin/Debug/SavedItems.json
#	ItemWatcher2/bin/Debug/rareWatcheItems.json

+++ ItemWatcher2/StartupForm.cs
@@ -62,7 +62,7 @@ namespace ItemWatcher2
 
                 
             config.alch_ratio = Convert.ToDecimal(this.txtAlch.Text);
-            config.exalt_ratio = Convert.ToInt32(this.txtExalt.Text);
+            config.exalt_ratio = Convert.ToDecimal(this.txtExalt.Text);
             config.fusing_ratio = Convert.ToDecimal(this.txtFuse.Text);
 
             config.profit_percent = Convert.ToDecimal(this.txtProfitPercent.Text);
